feat: use websocket header to pass the auth token
